+++ tests/testthat/test-layer_pointcloud.R
@@ -17,32 +17,32 @@ test_that("add_pointcloud accepts multiple objects", {
 	sf$elev <- 12345
 	sf <- sf::st_as_sf( sf, coords = c("lon", "lat", "elev") )
 	p <- add_pointcloud(map = m, data = sf)
-	expect_equal( as.character( p$x$calls[[1]]$args[[1]] ), geo )
+	expect_equal( as.character( p$x$calls[[1]]$args[[2]] ), geo )
 
 	## sfencoded
 	enc <- googlePolylines::encode( sf )
 	enc$z <- 123
 	p <- add_pointcloud( map = m, data = enc, elevation = "z" )
-	expect_equal( as.character( p$x$calls[[1]]$args[[1]] ), poly1 )
+	expect_equal( as.character( p$x$calls[[1]]$args[[2]] ), poly1 )
 
 	## sfencodedLite
 	enc <- googlePolylines::encode( sf, strip = T )
 	enc$z <- 123
 	p <- add_pointcloud( map = m, data = enc, elevation = "z" )
-	expect_equal( as.character( p$x$calls[[1]]$args[[1]] ), poly1 )
+	expect_equal( as.character( p$x$calls[[1]]$args[[2]] ), poly1 )
 
 	## data.frame with polyline
 	df <- as.data.frame( enc )
 	df$geometry <- unlist( df$geometry )
 	p <- add_pointcloud( map = m, data = df, elevation = "z", polyline = "geometry")
-	expect_equal( as.character( p$x$calls[[1]]$args[[1]] ), poly2 )
+	expect_equal( as.character( p$x$calls[[1]]$args[[2]] ), poly2 )
 
 	## data.frame
 	df <- capitals[1, ]
 	df$z <- 12345
 	geo <- '[{"type":"Feature","properties":{"fill_colour":"#440154FF"},"geometry":{"geometry":{"type":"Point","coordinates":[69.11,34.28,12345.0]}}}]'
 	p <- add_pointcloud( map = m, data = df, lon = "lon", lat = "lat", elevation = "z" )
-	expect_equal( as.character( p$x$calls[[1]]$args[[1]] ), geo )
+	expect_equal( as.character( p$x$calls[[1]]$args[[2]] ), geo )
 
 })