+++ tests/testthat/test-layer_polygon.R
@@ -17,7 +17,7 @@ test_that("add_polygon accepts multiple objects", {
 
 	sf <- spatialwidget::widget_melbourne[ spatialwidget::widget_melbourne$SA2_NAME == "South Yarra - West", ]
 	p <- add_polygon(map = m, data = sf)
-	expect_equal( as.character( p$x$calls[[1]]$args[[1]] ), geo )
+	expect_equal( as.character( p$x$calls[[1]]$args[[2]] ), geo )
 
 	## TODO( these tests, because some characters are being escaped in the polyline)
 	## sfencoded